Move Flutter Frames code into its own folder
