+++ packages/devtools_app/lib/src/screens/performance/panes/flutter_frames/flutter_frames_chart.dart
@@ -8,23 +8,23 @@ import 'dart:math' as math;
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
-import '../../analytics/analytics.dart' as ga;
-import '../../analytics/constants.dart' as analytics_constants;
-import '../../framework/scaffold.dart';
-import '../../primitives/auto_dispose_mixin.dart';
-import '../../primitives/utils.dart';
-import '../../shared/banner_messages.dart';
-import '../../shared/common_widgets.dart';
-import '../../shared/globals.dart';
-import '../../shared/theme.dart';
-import '../../shared/utils.dart';
-import '../../ui/colors.dart';
-import '../../ui/hover.dart';
-import '../../ui/utils.dart';
-import 'performance_controller.dart';
-import 'performance_model.dart';
-import 'performance_screen.dart';
-import 'performance_utils.dart';
+import '../../../../analytics/analytics.dart' as ga;
+import '../../../../analytics/constants.dart' as analytics_constants;
+import '../../../../framework/scaffold.dart';
+import '../../../../primitives/auto_dispose_mixin.dart';
+import '../../../../primitives/utils.dart';
+import '../../../../shared/banner_messages.dart';
+import '../../../../shared/common_widgets.dart';
+import '../../../../shared/globals.dart';
+import '../../../../shared/theme.dart';
+import '../../../../shared/utils.dart';
+import '../../../../ui/colors.dart';
+import '../../../../ui/hover.dart';
+import '../../../../ui/utils.dart';
+import '../../performance_controller.dart';
+import '../../performance_model.dart';
+import '../../performance_screen.dart';
+import '../../performance_utils.dart';
 
 // Turn this flag on to see when flutter frames are linked with timeline events.
 bool debugFrames = false;
