+++ packages/devtools_app/test/performance/performance_model_test.dart
@@ -9,7 +9,6 @@ import 'package:devtools_app/src/primitives/utils.dart';
 import 'package:devtools_app/src/screens/performance/panes/raster_stats/raster_stats_model.dart';
 import 'package:devtools_app/src/screens/performance/performance_model.dart';
 import 'package:devtools_app/src/screens/profiler/cpu_profile_model.dart';
-import 'package:devtools_app/src/service/service_manager.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import '../test_data/cpu_profile.dart';
@@ -402,47 +401,6 @@ void main() {
       expect(
         event.time.end!.inMicroseconds,
         asyncEndATrace.event.timestampMicros,
-      );
-    });
-  });
-
-  group('FlutterFrame', () {
-    test('shaderDuration', () {
-      expect(testFrame0.shaderDuration.inMicroseconds, equals(0));
-      expect(testFrame1.shaderDuration.inMicroseconds, equals(0));
-      expect(jankyFrame.shaderDuration.inMicroseconds, equals(0));
-      expect(jankyFrameUiOnly.shaderDuration.inMicroseconds, equals(0));
-      expect(jankyFrameRasterOnly.shaderDuration.inMicroseconds, equals(0));
-      expect(
-        testFrameWithShaderJank.shaderDuration.inMicroseconds,
-        equals(50000),
-      );
-      expect(
-        testFrameWithSubtleShaderJank.shaderDuration.inMicroseconds,
-        equals(4000),
-      );
-    });
-
-    test('hasShaderTime', () {
-      expect(testFrame0.hasShaderTime, isFalse);
-      expect(testFrame1.hasShaderTime, isFalse);
-      expect(jankyFrame.hasShaderTime, isFalse);
-      expect(jankyFrameUiOnly.hasShaderTime, isFalse);
-      expect(jankyFrameRasterOnly.hasShaderTime, isFalse);
-      expect(testFrameWithShaderJank.hasShaderTime, isTrue);
-      expect(testFrameWithSubtleShaderJank.hasShaderTime, isTrue);
-    });
-
-    test('hasShaderJank', () {
-      expect(testFrame0.hasShaderJank(defaultRefreshRate), isFalse);
-      expect(testFrame1.hasShaderJank(defaultRefreshRate), isFalse);
-      expect(jankyFrame.hasShaderJank(defaultRefreshRate), isFalse);
-      expect(jankyFrameUiOnly.hasShaderJank(defaultRefreshRate), isFalse);
-      expect(jankyFrameRasterOnly.hasShaderJank(defaultRefreshRate), isFalse);
-      expect(testFrameWithShaderJank.hasShaderJank(defaultRefreshRate), isTrue);
-      expect(
-        testFrameWithSubtleShaderJank.hasShaderJank(defaultRefreshRate),
-        isFalse,
       );
     });
   });
